Add CortexM3 Timer
