+++ qiling/hw/timer/cm3_systick.py
@@ -1,0 +1,12 @@
+#!/usr/bin/env python3
+# 
+# Cross Platform and Multi Architecture Advanced Binary Emulation Framework
+#
+
+import ctypes
+
+from qiling.hw.timer.cm_systick import CortexMSysTick
+
+
+class CortexM3SysTick(CortexMSysTick):
+    pass
